+++ src/config.json
@@ -1,6 +1,8 @@
 {
   "APP_NAME": "ArchiGuesser",
   "URL": "https://archiguesser.com",
+  "STARTING_DATE": [2022, 6, 1], 
+  "_STARTING_DATE": "// Here the first day was July (6) the 2nd (1) of 2022",
   "ITEMS_PER_LEVEL": 5,
   "MAX_LEVEL": 5,
   "ANONYMOUS_RETRIES": 3,
